Move TM2 datarequest to its own package
This breaks up a 1200-line file, into one file per endpoint.
