+++ traffic_monitor_golang/common/util/num.go
@@ -1,5 +1,7 @@
 package util
 
+const MillisecondsPerNanosecond = int64(1000000)
+
 // ToNumeric returns a float for any numeric type, and false if the interface does not hold a numeric type.
 // This allows converting unknown numeric types (for example, from JSON) in a single line
 // TODO try to parse string stats as numbers?
