+++ traffic_monitor_golang/traffic_monitor/datareq/monitorconfig.go
@@ -1,0 +1,11 @@
+package datareq
+
+import (
+	"encoding/json"
+
+	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/traffic_monitor/threadsafe"
+)
+
+func srvMonitorConfig(mcThs threadsafe.TrafficMonitorConfigMap) ([]byte, error) {
+	return json.Marshal(mcThs.Get())
+}
