+++ traffic_monitor_golang/traffic_monitor/threadsafe/monitorconfig.go
@@ -1,0 +1,62 @@
+package threadsafe
+
+import (
+	"sync"
+
+	to "github.com/apache/incubator-trafficcontrol/traffic_ops/client"
+)
+
+// CopyTrafficMonitorConfigMap returns a deep copy of the given TrafficMonitorConfigMap
+func CopyTrafficMonitorConfigMap(a *to.TrafficMonitorConfigMap) to.TrafficMonitorConfigMap {
+	b := to.TrafficMonitorConfigMap{}
+	b.TrafficServer = map[string]to.TrafficServer{}
+	b.CacheGroup = map[string]to.TMCacheGroup{}
+	b.Config = map[string]interface{}{}
+	b.TrafficMonitor = map[string]to.TrafficMonitor{}
+	b.DeliveryService = map[string]to.TMDeliveryService{}
+	b.Profile = map[string]to.TMProfile{}
+	for k, v := range a.TrafficServer {
+		b.TrafficServer[k] = v
+	}
+	for k, v := range a.CacheGroup {
+		b.CacheGroup[k] = v
+	}
+	for k, v := range a.Config {
+		b.Config[k] = v
+	}
+	for k, v := range a.TrafficMonitor {
+		b.TrafficMonitor[k] = v
+	}
+	for k, v := range a.DeliveryService {
+		b.DeliveryService[k] = v
+	}
+	for k, v := range a.Profile {
+		b.Profile[k] = v
+	}
+	return b
+}
+
+// TrafficMonitorConfigMapThreadsafe encapsulates a TrafficMonitorConfigMap safe for multiple readers and a single writer.
+type TrafficMonitorConfigMap struct {
+	monitorConfig *to.TrafficMonitorConfigMap
+	m             *sync.RWMutex
+}
+
+// NewTrafficMonitorConfigMap returns an encapsulated TrafficMonitorConfigMap safe for multiple readers and a single writer.
+func NewTrafficMonitorConfigMap() TrafficMonitorConfigMap {
+	return TrafficMonitorConfigMap{monitorConfig: &to.TrafficMonitorConfigMap{}, m: &sync.RWMutex{}}
+}
+
+// Get returns the TrafficMonitorConfigMap. Callers MUST NOT modify, it is not threadsafe for mutation. If mutation is necessary, call CopyTrafficMonitorConfigMap().
+func (t *TrafficMonitorConfigMap) Get() to.TrafficMonitorConfigMap {
+	t.m.RLock()
+	defer t.m.RUnlock()
+	return *t.monitorConfig
+}
+
+// Set sets the TrafficMonitorConfigMap. This is only safe for one writer. This MUST NOT be called by multiple threads.
+func (t *TrafficMonitorConfigMap) Set(c to.TrafficMonitorConfigMap) {
+	t.m.Lock()
+	*t.monitorConfig = c
+	t.m.Unlock()
+}
