+++ traffic_monitor_golang/traffic_monitor/traffic_monitor.go
@@ -20,17 +20,12 @@ package main
  */
 
 import (
-	"bytes"
 	"flag"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"math"
 	"os"
-	"os/exec"
-	"path/filepath"
 	"runtime"
-	"time"
 
 	_ "github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/instrumentation"
 	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
@@ -44,46 +39,6 @@ var GitRevision = "No Git Revision Specified. Please build with '-X main.GitRevi
 
 // BuildTimestamp is the time the app was built. The app SHOULD always be built with this set via the `-X` flag.
 var BuildTimestamp = "No Build Timestamp Specified. Please build with '-X main.BuildTimestamp=`date +'%Y-%M-%dT%H:%M:%S'`"
-
-// getHostNameWithoutDomain returns the machine hostname, without domain information.
-// Modified from http://stackoverflow.com/a/34331660/292623
-func getHostNameWithoutDomain() (string, error) {
-	cmd := exec.Command("/bin/hostname", "-s")
-	var out bytes.Buffer
-	cmd.Stdout = &out
-	err := cmd.Run()
-	if err != nil {
-		return "", err
-	}
-	hostname := out.String()
-	if len(hostname) < 1 {
-		return "", fmt.Errorf("OS returned empty hostname")
-	}
-	hostname = hostname[:len(hostname)-1] // removing EOL
-	return hostname, nil
-}
-
-// getStaticAppData returns app data available at start time.
-// This should be called immediately, as it includes calculating when the app was started.
-func getStaticAppData() (manager.StaticAppData, error) {
-	var d manager.StaticAppData
-	var err error
-	d.StartTime = time.Now()
-	d.GitRevision = GitRevision
-	d.FreeMemoryMB = math.MaxUint64 // TODO remove if/when nothing needs this
-	d.Version = Version
-	if d.WorkingDir, err = os.Getwd(); err != nil {
-		return manager.StaticAppData{}, err
-	}
-	d.Name = os.Args[0]
-	d.BuildTimestamp = BuildTimestamp
-	if d.Hostname, err = getHostNameWithoutDomain(); err != nil {
-		return manager.StaticAppData{}, err
-	}
-
-	d.UserAgent = fmt.Sprintf("%s/%s", filepath.Base(d.Name), d.Version)
-	return d, nil
-}
 
 func getLogWriter(location string) (io.Writer, error) {
 	switch location {
@@ -124,7 +79,7 @@ func getLogWriters(eventLoc, errLoc, warnLoc, infoLoc, debugLoc string) (io.Writ
 func main() {
 	runtime.GOMAXPROCS(runtime.NumCPU())
 
-	staticData, err := getStaticAppData()
+	staticData, err := config.GetStaticAppData(Version, GitRevision, BuildTimestamp)
 	if err != nil {
 		fmt.Printf("Error starting service: failed to get static app data: %v\n", err)
 		os.Exit(1)
